+++ CHANGELOG.markdown
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2020-09-04
+### Added
+-  Added a Syslog Audit Message Processor: The new Message Audit Processor is 
+based on the syslog-java-client library. This allows messages to be sent to a 
+syslog server. The supported protocols are UDP and TCP, the latter also with or 
+without SSL/TLS support. For more information refer to https://github.com/CloudBees-community/syslog-java-client
+
+## [1.2.0] - 2020-09-04
+### Added
+- Added support for Liferay 7.3 GA5
+
 ## [1.1.6] - 2020-09-03
 ### Added
 - Added support for Liferay 7.2 GA2